requirements.txt
pillow
requests
torch
torchvision
numpy==1.26.0